refactor: co-locate consumption export workflow ID helpers with their siblings
makeConsumptionExportWorkflowId/Prefix are naming helpers for the same
"consumption export" concept as buildConsumptionExportGcsPrefix,
buildConsumptionExportGcsPath and buildConsumptionExportCacheKey, and are
consumed from lib/api like those are. Move them next to their siblings in
activities/consumption_export.ts instead of the temporal-only helpers.ts,
which now only holds the agent-message-analytics ID builder used solely
by client.ts.
